MSPApp: capturing images upon motion working STMApp/STMLoader: updates for Toastbox changes

+++ Tools/MDCUtil/main.cpp
@@ -63,7 +63,7 @@ struct Args {
     } MSPLoad = {};
     
     struct {
-        uint8_t idx = 0;
+        uint32_t idx = 0;
         std::string filePath;
     } SDImgRead = {};
